move drawing-layer and drawing-object
