+++ src/plugins/drawing-tool/components/drawing-object.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { DrawingObjectDataType, Point } from "../../../plugins/drawing-tool/model/drawing-objects";
+import { DrawingObjectDataType, Point } from "../model/drawing-objects";
 import { uniqueId } from "../../../utilities/js-utils";
 import { DrawingLayerView } from "./drawing-layer";
