+++ src/plugins/drawing-tool/components/drawing-tool.tsx
@@ -2,7 +2,7 @@ import classNames from "classnames";
 import React, { useEffect } from "react";
 import { IToolTileProps } from "../../../components/tools/tool-tile";
 import { ToolbarView } from "./drawing-toolbar";
-import { DrawingLayerView } from "../../../components/tools/drawing-tool/drawing-layer";
+import { DrawingLayerView } from "./drawing-layer";
 import { useToolbarToolApi } from "../../../components/tools/hooks/use-toolbar-tool-api";
 import { DrawingContentModelType } from "../model/drawing-content";
 import { useCurrent } from "../../../hooks/use-current";
